set the id types to 8-byte unsigned integers as discussed on the call

+++ Include/MonarchTypes.hpp
@@ -50,8 +50,8 @@ static const FormatModeType sFormatSingle = 0;
 static const FormatModeType sFormatMultiSeparate = 1;
 static const FormatModeType sFormatMultiInterleaved = 2;
 
-typedef uint32_t AcquisitionIdType; // 4 bytes
-typedef uint32_t RecordIdType; // 4 bytes
+typedef uint64_t AcquisitionIdType; // 8 bytes
+typedef uint64_t RecordIdType; // 8 bytes
 typedef uint64_t TimeType; // 8 bytes
 typedef uint8_t DataType; // 1 byte
